+++ y/iterator.go
@@ -27,8 +27,8 @@ type ValueStruct struct {
 }
 
 // EncodedSize is the size of the ValueStruct when encoded
-func (v *ValueStruct) EncodedSize() uint16 {
-	return uint16(len(v.Value) + len(v.UserMeta) + 2) // meta
+func (v *ValueStruct) EncodedSize() uint32 {
+	return uint32(len(v.Value) + len(v.UserMeta) + 2) // meta
 }
 
 // Decode uses the length of the slice to infer the length of the Value field.
